Remove the android namespace from attributes in AndroidManifest.xml
Some apk files contain AndroidManifest.xml,
in which the namespace with value
http://schemas.android.com/apk/res/android
undefined, despite the fact that Android Studio
does not allow you to do this.
Then the manifest is incorrectly parsed.
Therefore, the namespace requirement has been removed.

NOTE: in the original repository, this solution
will not work starts with version 1.0.4.